move actions around to make space for ac002
