+++ pkg/action/cl001/ac002/explainer.go
@@ -5,13 +5,5 @@ import (
 )
 
 func (e *Explainer) explain(ctx context.Context) (string, error) {
-	s := `
-Check if the desired amount of Tenant Cluster master nodes are up and ready.
-
-	* Fetch all G8sControlPlane CRs spec.replicas so that we know how many masters the Tenant Cluster is supposed to have.
-	* Fetch the Tenant Cluster master nodes.
-	* Compare the current and desired amount of master nodes.
-	`
-
-	return s, nil
+	return "", nil
 }
